+++ common/src/main/java/smartin/miapi/item/modular/items/ItemStackSensitive.java
@@ -1,0 +1,7 @@
+package smartin.miapi.item.modular.items;
+
+import net.minecraft.item.ItemStack;
+
+public interface ItemStackSensitive {
+    void lastItemStack(ItemStack itemStack);
+}