+++ common/src/main/java/smartin/miapi/item/modular/items/ModularAxe.java
@@ -47,7 +47,7 @@ public class ModularAxe extends AxeItem implements PlatformModularItemMethods, M
     }
 
     @Override
-    public void setToolMaterial(ToolMaterial toolMaterial) {
+    public void lastItemStack(ToolMaterial toolMaterial) {
         this.currentFakeToolmaterial = toolMaterial;
     }