+++ common/src/main/java/smartin/miapi/item/modular/items/ModularSetableToolMaterial.java
@@ -4,10 +4,10 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
 import smartin.miapi.modules.properties.mining.MiningLevelProperty;
 
-public interface ModularSetableToolMaterial {
-    void setToolMaterial(ToolMaterial toolMaterial);
+public interface ModularSetableToolMaterial extends ItemStackSensitive{
+    void lastItemStack(ToolMaterial toolMaterial);
 
-    default void setToolMaterial(ItemStack itemStack) {
-        setToolMaterial(MiningLevelProperty.getFakeToolMaterial(itemStack));
+    default void lastItemStack(ItemStack itemStack) {
+        lastItemStack(MiningLevelProperty.getFakeToolMaterial(itemStack));
     }
 }