+++ common/src/main/java/smartin/miapi/item/modular/items/ModularSword.java
@@ -52,7 +52,7 @@ public class ModularSword extends SwordItem implements PlatformModularItemMethod
     }
 
     @Override
-    public void setToolMaterial(ToolMaterial toolMaterial) {
+    public void lastItemStack(ToolMaterial toolMaterial) {
         this.currentFakeToolmaterial = toolMaterial;
     }