+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - fixed limit not being applied correctly, resulting in too few materials being generated
 - changed ability logic to allow other mods to take right click action if truly modular doesnt set one
 - added "copy_item" ability, trying to copy another items ability
+- improved armor compatibility